Fix the ajabase build on Windows and fix some warnings

+++ ajalibraries/ajabase/system/info.cpp
@@ -112,7 +112,7 @@ AJASystemInfo::ToString(int maxLength)
         if (retLabel == AJA_STATUS_SUCCESS && retValue == AJA_STATUS_SUCCESS)
         {
             if ((int)label.length() > longestLabelLen)
-                longestLabelLen = label.length();
+                longestLabelLen = (int)label.length();
         }
     }
